smoke-tests/88-pre-existing-otel/fixtures/partial-otel-app/requirements.txt
flask>=3.0.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-http>=1.22.0
opentelemetry-instrumentation-flask>=0.43b0
